Updated WebApi Client.
Added MSBuildTasks/NUnit runner for build automation.

Moved service tests that use live access to Natatmo into Integration tests project.

Added sample data to test project to test the response objects map correctly to response json.

Moved some of the mapping out from services to dedicated mappers for better re-use and testability.

Changed DateTimes to nullables as Netatmo returns some null dates.

Improved support for rain gauges that now include different additional data.

Added support for calculating distance from the user of the various sensors.

Added support for adjustable perimiter around the user for getting public data (e.g. 10k box, 2k box etc).

Fixed bug where it appears Netatmo return place as [long,lat] rather than lat,long.

Removed a number of properties from User model that Netatmo no longer appear to be returning.

Introduced HttpWrapper to allow faking of http get call for testing.

+++ NetatmoBot/Services/PublicDataModels/Place.cs
@@ -2,8 +2,9 @@ namespace NetatmoBot.Services.PublicDataModels
 {
     public class Place
     {
-        public decimal[] location { get; set; }
-        public decimal altitude { get; set; }
+        // Appears to be [lon,lat]
+        public double[] location { get; set; }
+        public double altitude { get; set; }
         public string timezone { get; set; }
 
         // Present in module list.
